Update heading styles and add hover effect
The commit updates the CSS and Sass files to add a new hover effect on headings. The update includes changes to the background-clip, color, letter-spacing, and transition properties of the .heading-secondary class in main.css as well as changes to the same properties in _typography.scss. Additionally, a new transform property is added for the :hover pseudo-class of both classes.