Refactor composition photo styles for responsive design
This commit refactors the CSS styles for the composition photo element to make it more responsive. The width of the images is now set as a percentage of their container, and the box-shadow and z-index are increased when hovering over an image to give the impression that it is closer to the user.