Refactor composition photo hover effect
The commit refactors the composition photo hover effect by adding a translateY transform to move the image up and down. The box-shadow is also increased and made darker to give the impression that the image is closer to the user. Additionally, there is a minor change in _composition.scss file where an extra closing parenthesis was added.